lmb-414: move fetching the sub row data to the router
