+++ app/components/mandatarissen/persoon-table-row.js
@@ -1,61 +1,34 @@
 import Component from '@glimmer/component';
 
 import { tracked } from '@glimmer/tracking';
-import { A } from '@ember/array';
 import { action } from '@ember/object';
-
-import { foldMandatarisses } from 'frontend-lmb/utils/fold-mandatarisses';
 
 export default class MandatarissenPersoonTableRowComponent extends Component {
   @tracked isSubRowOpen;
-  @tracked subRows = A();
 
   @action
   openCloseSubRows() {
     this.isSubRowOpen = !this.isSubRowOpen;
   }
 
-  @action
-  async getSubRowData() {
-    const foldedMandatarissen = await foldMandatarisses(
-      null,
-      this.args.mandatarissen
-    );
+  get firstRow() {
+    return this.args.mandatarissen.find((value) => value.isSubRow === false);
+  }
 
-    await Promise.all(
-      foldedMandatarissen.map(async (foldedMandataris) => {
-        const mandataris = foldedMandataris.mandataris;
-        const lidmaatschap = await mandataris.heeftLidmaatschap;
-        const mandaat = await mandataris.bekleedt;
-        const bestuursfunctie = await mandaat.bestuursfunctie;
-        const bestuursorganenInTijd = await mandaat.bevatIn;
-        let bestuursorgaan = null;
-        let fractieLabel = null;
+  get subRows() {
+    return this.args.mandatarissen.map((value) => value.rowData);
+  }
 
-        if (bestuursorganenInTijd.length >= 1) {
-          const bestuursorgaanInTijd = bestuursorganenInTijd.at(0);
-          bestuursorgaan = await bestuursorgaanInTijd.isTijdsspecialisatieVan;
-        }
-        if (!lidmaatschap) {
-          fractieLabel = 'Niet beschikbaar';
-        } else {
-          fractieLabel = (await lidmaatschap.binnenFractie)?.naam;
-        }
+  get mandatarissen() {
+    return this.args.mandatarissen.map((value) => value.mandataris);
+  }
 
-        this.subRows.pushObject({
-          mandataris: mandataris,
-          fractie: fractieLabel,
-          bestuursorgaan: {
-            label: bestuursorgaan?.naam,
-            routeModelId: bestuursorgaan?.id,
-          },
-          mandaat: {
-            label: bestuursfunctie.label,
-            routeModelIds: [this.args.persoon.id, mandataris.id],
-          },
-        });
-      })
-    );
+  get hasOnlyOneMandaat() {
+    return this.mandatarissen.length === 1;
+  }
+
+  get showCombinedInfo() {
+    return this.mandatarissen.length > 1 && !!this.isSubRowOpen === false;
   }
 
   get persoonDetailRoute() {
@@ -64,9 +37,5 @@ export default class MandatarissenPersoonTableRowComponent extends Component {
 
   get iconSubRowOpen() {
     return this.isSubRowOpen ? 'nav-up' : 'nav-down';
-  }
-
-  get firstSubRow() {
-    return this.subRows?.firstObject;
   }
 }
